feat: Add ProductCategoryFactory for generating product category instances
This commit adds the ProductCategoryFactory file, which is responsible for generating instances of the ProductCategory model for testing purposes. The factory defines the default state of the model, including generating a unique name and a slug based on the name using the `Str::slug` method.

This factory will be used to create test data for the product category functionality in the application.